requirements.txt
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.1.0
scikit_image==0.19.2
scikit_learn==1.0.2
segmentation_models_pytorch==0.3.2
timm==0.6.12
torch==1.9.0+cu111
torchmetrics==1.1.2
torchvision==0.10.0+cu111
